backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
gunicorn
whitenoise
dj-database-url
python-decouple
django-environ
drf-yasg
requests
polyline